+++ blockindex/indexpb/liquidstaking_bucket.proto
@@ -1,0 +1,24 @@
+// Copyright (c) 2019 IoTeX
+// This source code is provided 'as is' and no warranties are given as to title or non-infringement, merchantability
+// or fitness for purpose and, to the extent permitted by law, all liability for your use of the code is disclaimed.
+// This source code is governed by Apache License 2.0 that can be found in the LICENSE file.
+
+// To compile the proto, run:
+//      protoc --go_out=plugins=grpc:. *.proto
+syntax = "proto3";
+package indexpb;
+option go_package = "github.com/iotexproject/iotex-core/blockindex/indexpb";
+import "google/protobuf/timestamp.proto";
+
+message BucketType {
+    string amount = 1;
+    uint64 duration = 2;
+    google.protobuf.Timestamp activatedAt = 3;
+}
+
+message BucketInfo {
+    uint64 typeIndex = 1;
+    google.protobuf.Timestamp unlockedAt = 2;
+    google.protobuf.Timestamp unstakedAt = 3;
+    string delegate = 4;
+}
